% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSentimentWordCloud.R
\name{CreateSentimentWordCloud}
\alias{CreateSentimentWordCloud}
\title{Create a Word Cloud for Sentiment Analysis}
\usage{
CreateSentimentWordCloud(sentiment_data, tokenized)
}
\arguments{
\item{sentiment_data}{ONLY PASS THIS PARAM!!! A tibble of sentiment analysis results (either from \code{SentimentAnalysis} or \code{SentimentTokenized}).}

\item{tokenized}{A \code{TokenizedText} object used to create the word cloud.}
}
\value{
A word cloud visualization.
}
\description{
Generates a word cloud with color-coding based on sentiment (positive = green, negative = red).
}
