% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateWordCloud.R
\name{CreateWordCloud}
\alias{CreateWordCloud}
\title{Create Word Cloud}
\usage{
CreateWordCloud(text)
}
\arguments{
\item{text}{A character vector or an object of class \code{TokenizedText}.}
}
\value{
A word cloud plot showing word frequencies.
}
\description{
Generates a word cloud from raw or tokenized text.
}
\examples{
CreateWordCloud("This is an example tweet! Tweet tweet!")
}
