% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SentimentAnalysis.R
\name{SentimentAnalysis}
\alias{SentimentAnalysis}
\title{Sentiment Analysis for Raw Text}
\usage{
SentimentAnalysis(text)
}
\arguments{
\item{text}{A character vector containing the raw text.}
}
\value{
A tibble with words, sentiment classification, and sentiment score.
}
\description{
Analyzes the sentiment of raw text, returning the sentiment (positive/negative) and sentiment score.
Tokenizes the text internally and uses the \code{bing} lexicon to classify words.
}
\examples{
sentiment <- SentimentAnalysis("This is wonderful and amazing but also terrible.")
}
