% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SentimentTokenized.R
\name{SentimentTokenized}
\alias{SentimentTokenized}
\title{Sentiment Analysis for Tokenized Text}
\usage{
SentimentTokenized(tokenized)
}
\arguments{
\item{tokenized}{A \code{TokenizedText} object from the \code{TokenizeText} function.}
}
\value{
A tibble with words, sentiment classification, sentiment score, and frequency of each word.
}
\description{
Analyzes the sentiment of tokenized text, returning the sentiment (positive/negative) and sentiment score.
Uses the \code{bing} lexicon to classify words.
}
\examples{
tokenized <- TokenizeText("This is wonderful and amazing but also terrible.")
sentiment <- SentimentTokenized(tokenized)
}
