% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TokenizeText.R
\name{TokenizeText}
\alias{TokenizeText}
\title{Tokenize Text}
\usage{
TokenizeText(text)
}
\arguments{
\item{text}{A character vector containing the text to tokenize.}
}
\value{
An object of class \code{TokenizedText} (a data frame with \code{word} and \code{count} columns).
}
\description{
Tokenize Text
}
\examples{
TokenizeText(c("This is an example tweet!", "Tweet tweet!"))
}
